Use Unexceptional and export new API
Always use UTCTime, no reason for a date parser to return Integer and
then export an extra conversion function everyone is forced to use.

Effectful approxidate is now called approxidateIO following closer to
normal naming conventions.  Though this is maybe a strangish change
since it now returns (Unexceptional m) -- but IO is a member of that at
least.